pyproject.toml
[tool.poetry]
name = "main.py"
version = "0.1.0"
description = ""
authors = ["JasonxDev <jasonxdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
